Break project into modules
