+++ src/js/.jshintrc
@@ -1,5 +1,8 @@
 {
     "undef": true,
     "unused": true,
-    "browser": true
+    "browser": true,
+    "esnext": true,
+
+    "predef": [ "moment" ]
 }
